requirements.txt
feedparser==6.0.10
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
schedule==1.2.1